reorganized gatsby components
