+++ components/src/components/Element/Button/index.js
@@ -1,3 +1,3 @@
 export { Button } from './Button'
 export { LinkButton } from './LinkButton'
-export {IconButton} from './IconButton'
+export { IconButton} from './IconButton'
