+++ components/src/gatsby-components/image.js
