+++ components/src/layouts/layout.js
@@ -1,7 +1,7 @@
 import React from "react"
 import PropTypes from "prop-types"
 import { StaticQuery, graphql } from "gatsby"
-import Header from "./header"
+import Header from "../gatsby-components/header"
 import "./layout.css"
 import {ThemeProvider} from 'styled-components'
 import {Gray as theme } from '../themes/Gray'
