+++ components/src/pages/index.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import { Link } from 'gatsby'
 
-import Layout from '../components/layout'
-import Image from '../components/image'
-import SEO from '../components/seo'
+import Layout from '../layouts/layout'
+import Image from '../gatsby-components/image'
+import SEO from '../gatsby-components/seo'
 import { Button,IconButton, LinkButton,Section, Address } from '../components/Element'
 import {Comment} from 'styled-icons/octicons'
