+++ components/src/pages/page-2.js
@@ -1,8 +1,8 @@
 import React from "react"
 import { Link } from "gatsby"
 
-import Layout from "../components/layout"
-import SEO from "../components/seo"
+import Layout from "../layouts/layout"
+import SEO from "../gatsby-components/seo"
 
 const SecondPage = () => (
   <Layout>
